# swagger.yaml
openapi: 3.0.3
info:
  title: Dice Engage — Omnichannel SaaS Platform API
  description: |
    Production-ready, multi-tenant omnichannel marketing SaaS API.
    Supports Email, SMS, WhatsApp, Mobile Push, Web Push, RCS, and Webhooks
    with tenant isolation, billing, analytics, and admin management.
  version: 1.0.0
  contact:
    name: Dice Engage Team
servers:
  - url: http://localhost:3001
    description: Local development
  - url: https://api.diceengage.com
    description: Production
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    ApiKeyAuth:
      type: apiKey
      name: Authorization
      in: header
    TenantContext:
      type: apiKey
      name: x-tenant-id
      in: header
  schemas:
    Error:
      type: object
      required: [message]
      properties:
        message:
          type: string
    TenantResource:
      type: object
      required: [id, name, planType, status, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        planType:
          type: string
          enum: [Starter, Growth, Enterprise]
        status:
          type: string
          enum: [Active, Suspended, Cancelled]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    UpsertTenantResource:
      type: object
      required: [name]
      properties:
        name:
          type: string
        planType:
          type: string
          enum: [Starter, Growth, Enterprise]
    BrandResource:
      type: object
      required: [id, tenantId, name, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        tenantId:
          type: string
          format: uuid
        name:
          type: string
        timezone:
          type: string
        senderConfigJson:
          type: object
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    TenantWithUsage:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        planType:
          type: string
        status:
          type: string
        workspaceCount:
          type: integer
        currentMonthMessages:
          type: integer
        currentMonthCostMicros:
          type: integer
        createdAt:
          type: string
          format: date-time
...
security:
  - BearerAuth: []
paths:
...
  /api/admin/tenants:
    post:
      tags: [Tenants]
      summary: Create a new tenant
      operationId: createTenant
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpsertTenantResource"
      responses:
        "201":
          description: Tenant created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TenantResource"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    get:
      tags: [Tenants]
      summary: List all tenants
      operationId: listTenants
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [Active, Suspended, Cancelled]
      responses:
        "200":
          description: Tenant list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TenantResource"
  /api/admin/tenants/{tenantId}:
    get:
      tags: [Tenants]
      summary: Get a tenant by ID
      operationId: getTenant
      parameters:
        - name: tenantId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Tenant found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TenantResource"
        "404":
          description: Tenant not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    put:
      tags: [Tenants]
      summary: Update a tenant
      operationId: updateTenant
      parameters:
        - name: tenantId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpsertTenantResource"
      responses:
        "200":
          description: Tenant updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TenantResource"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    delete:
      tags: [Tenants]
      summary: Delete a tenant
      operationId: deleteTenant
      parameters:
        - name: tenantId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Tenant deleted
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/admin/tenants/{tenantId}/suspend:
    post:
      tags: [Tenants]
      summary: Suspend a tenant
      operationId: suspendTenant
      parameters:
        - name: tenantId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Tenant suspended
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TenantResource"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/admin/tenants/{tenantId}/activate:
    post:
      tags: [Tenants]
      summary: Activate a suspended tenant
      operationId: activateTenant
      parameters:
        - name: tenantId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Tenant activated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TenantResource"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/admin/tenants/{tenantId}/workspaces:
    post:
      tags: [Tenants]
      summary: Assign a workspace to a tenant
      operationId: assignWorkspaceToTenant
      parameters:
        - name: tenantId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [workspaceId]
              properties:
                workspaceId:
                  type: string
                  format: uuid
                brandId:
                  type: string
                  format: uuid
      responses:
        "204":
          description: Workspace assigned
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  # ═══════════════════════════════════════════════════════════════════
  # BRAND MANAGEMENT — /api/admin/tenants/{tenantId}/brands
  # ═══════════════════════════════════════════════════════════════════
  /api/admin/tenants/{tenantId}/brands:
    post:
      tags: [Brands]
      summary: Create a brand under a tenant
      operationId: createBrand
      parameters:
        - name: tenantId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                timezone:
                  type: string
                senderConfigJson:
                  type: object
      responses:
        "201":
          description: Brand created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BrandResource"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    get:
      tags: [Brands]
      summary: List brands for a tenant
      operationId: listBrands
      parameters:
        - name: tenantId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Brand list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BrandResource"
  /api/admin/tenants/{tenantId}/brands/{brandId}:
    put:
      tags: [Brands]
      summary: Update a brand
      operationId: updateBrand
      parameters:
        - name: tenantId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: brandId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                timezone:
                  type: string
                senderConfigJson:
                  type: object
      responses:
        "200":
          description: Brand updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BrandResource"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    delete:
      tags: [Brands]
      summary: Delete a brand
      operationId: deleteBrand
      parameters:
        - name: tenantId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: brandId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Brand deleted
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  # ═══════════════════════════════════════════════════════════════════
  # BILLING & ANALYTICS — /api/admin/billing
  # ═══════════════════════════════════════════════════════════════════
...
  /api/admin/panel/tenants:
    get:
      tags: [Admin Panel]
      summary: List tenants with usage stats
      operationId: listTenantsWithUsage
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [Active, Suspended, Cancelled]
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: Paginated tenant list with usage
          content:
            application/json:
              schema:
                type: object
                properties:
                  tenants:
                    type: array
                    items:
                      $ref: "#/components/schemas/TenantWithUsage"
                  total:
                    type: integer
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /api/admin/panel/tenants/{tenantId}/suspend:
    post:
      tags: [Admin Panel]
      summary: Suspend a tenant (admin panel)
      operationId: adminSuspendTenant
      parameters:
        - name: tenantId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Tenant suspended
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  status:
                    type: string
        "404":
          description: Tenant not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/admin/panel/tenants/{tenantId}/activate:
    post:
      tags: [Admin Panel]
      summary: Activate a tenant (admin panel)
      operationId: adminActivateTenant
      parameters:
        - name: tenantId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Tenant activated
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  status:
                    type: string
        "404":
          description: Tenant not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...